Minimum working example for sound
Only implements the tone channel (not the sweep) to test whether this
implementation works. Adding the sweep channel shouldn't be too much
extra work, but this is getting to be a lot of work to have only on my
local machine.

This sound implementation integrates reasonably well (not too many
changes) with the existing framework. As the motherboard steps the cpu
through clocks, it adds to a counter in the Sound module to signal how
many cycles have passed. When an instruction arrives that might affect
the state of the sound module (setting or reading a sound register) or
when a frame is finished and rendered, the Sound module is called to
run its machine until it's current, and it generates any audio samples
for that period of time and sends them off to sdl2/hardware. This
operation is blocking.

Doing it this way, I'm assuming that computation happens fast enough
(or the frame is ready fast enough?) that the sound can come out fast
enough that it won't fall behind or leave gaps. At most, it should be
a frame behind the cpu, which is probably what matches the display
anyway. Right now, though, it lags a couple seconds behind for some
reason. I'm not sure what to do about that, but it will probably be
fixed at the same time as when I figure out how to handle running
faster than 100% speed.

This is a Python-only implementation so far, which works fine and
keeps up even on my netbook when the rest of PyBoy is cythonized.
Cython compatibility can/will be added later, but for development's
sake I'm just doing it in Python, and it's really nice that it works
that way tbh.