Added sweep channel, but it's not quite working.
Something's not quite right with the sweep channel--actually, I'm not
sure if it's right at all. Both the Game Freak logo and the Jigglypuff
jumps in the intro don't sound like I expect them to, even with the
two missing channels left. It's hard to find other examples of the
sweep function to test with.

I also fixed a mistake where I forgot to implement the length timer
right/at all. I'm not 100% sure it's correct now either, because the
specs are a little sparse on it, but I'm not entirely sure how it
could be wrong.

The out-of-sync problem is still there, and it's bad enough that I
think I need to fix that first. It may be that the sample rate is
lower than what sdl2 reports, or (less likely) that gaps in the sound
between frames are adding up since we're not calculating ahead, but
I'd think that would be audible.