Fixed the sweep channel problem
The sweep channel didn't sound right; it turned out that it was
because I was triggering in setreg() before finishing the register
writes, so the upper 3 bits of the period were being discarded.

Still not sure what to do about synchronization. The emulator doesn't
run the 0x400000 clocks per second I expected, so that's a problem,
but it's not enough to explain the size of the desyncing anyway. I'll
probably just have to change how it's clocked entirely and accomodate
faster/slower emulation at the same time.