Minor changes for pre-commit compliance
Added per-file-ignores to .flake8 config so the full check can run but won't complain about the long comments in some of the example scripts.

Added autoenv directories to .gitignore because everyone's setup is different.

isort made some changes to the imports in a few of the files that don't get much attention.