Cartridge cythonized
* Removed options that generated warnings

`find -d` is for depth arguments, not finding directories
(`find -type d`), which I think is what was intended, but
also unnecessary.

`content_type` doesn't seem to do anything for `setup.py`

* Made changes necessary to enable cythonization of the cartridge module

As it turns out, there are a couple bugs in Cython that prevented
load_romfile in `cartridge.py` from producing working code. The
workflow required conversion from a list of `array.arrays` into a
memoryview object (actually, that might not have been valid, so not a
bug), and that failed in a way that caused deconstructor code to be
run on an unintitialized pointer, which caused a segfault. It may be
useful to address this one day, so I'm leaving the message here.

In this new version, the rombanks object is a shaped memoryview in
cython and a list of memoryview slices under PyPy. I tested all 3
configurations (including CPython) and they seem to work just fine.
It's a little unfortunate that the mbc's `rombanks` and `rambanks` are
now entirely different objects, but it seemed better to not change
both when `rambanks` worked just fine as is. Still, that's something
that could be considered if we ever rework that module as a whole.

* Cartridge() needs to be declared with a cpdef to be visible from the Python side

* Changed Cartridge to load_cartridge and fixed setup.py path