Windows native (#93)
* Removed unused SDL2 .pxd files. We won't need them.

* Support native Windows compilation with setup.py rather than make.

With these changes, PyBoy can be compiled with Microsoft's Visual
Studio build tools, instead of requiring mingw and using make, it is
invoked with `python setup.py build_ext -i` directly. Some work had to
go into finding and loading SDL2.dll and the header files, but it was
done in a way that is compatible with minGW/macOS/Linux as well. In
particular, `setup.py` can detect a minGW installation and have it run
`sdl2-config`, though it has to do this by finding and launching
msys2's `sh.exe`, which is a little messy.

As can be seen, the Makefile now does less of the SDL2 stuff, and
that's taken over by the setup.py. I have not yet tested whether
setup.py can also correctly run the install command, particularly
because whether or not it needs to run build_ext as part of that
process depends on whether PyPy is being used (vs Cython). I guess we
could just short-circuit the build_ext command if PyPy is detected,
but I'm not sure that's the best approach. But that comes later...

Also for convenience in Windows, `python setup.py clean -i` will
remove compiled files in the source tree, just like `make clean` does
now. It has the disadvantage that `setup.py` will cythonize the files
before running... but I'm not sure if it's possible to avoid that
without very invasive changes.

Oh, also I took out the stuff about including numpy headers. We don't
use `cimport numpy`, just regular import, so we don't need it.

The change to window_sdl2.py fixes a cast bug from Cython by just
avoiding it.

Some items were added to .gitignore for convenience on Windows.

With Python 3.8 out, I removed the version specification in the isort
section of .pre-commit-config.yaml. It still seems to run as always,
and import mechanics haven't changed between 3.7 and 3.8.

* Just consistency

* Minor fixes

* Fixed clean and added a fix to Cythons recompile-detection

* Added rewind.py to setup.py

Co-authored-by: Kristian Sims <kristian.sims@gmail.com>