Remaining channels to sound (#128)
* Minimum working example for sound

Only implements the tone channel (not the sweep) to test whether this
implementation works. Adding the sweep channel shouldn't be too much
extra work, but this is getting to be a lot of work to have only on my
local machine.

This sound implementation integrates reasonably well (not too many
changes) with the existing framework. As the motherboard steps the cpu
through clocks, it adds to a counter in the Sound module to signal how
many cycles have passed. When an instruction arrives that might affect
the state of the sound module (setting or reading a sound register) or
when a frame is finished and rendered, the Sound module is called to
run its machine until it's current, and it generates any audio samples
for that period of time and sends them off to sdl2/hardware. This
operation is blocking.

Doing it this way, I'm assuming that computation happens fast enough
(or the frame is ready fast enough?) that the sound can come out fast
enough that it won't fall behind or leave gaps. At most, it should be
a frame behind the cpu, which is probably what matches the display
anyway. Right now, though, it lags a couple seconds behind for some
reason. I'm not sure what to do about that, but it will probably be
fixed at the same time as when I figure out how to handle running
faster than 100% speed.

This is a Python-only implementation so far, which works fine and
keeps up even on my netbook when the rest of PyBoy is cythonized.
Cython compatibility can/will be added later, but for development's
sake I'm just doing it in Python, and it's really nice that it works
that way tbh.

* Added sweep channel, but it's not quite working.

Something's not quite right with the sweep channel--actually, I'm not
sure if it's right at all. Both the Game Freak logo and the Jigglypuff
jumps in the intro don't sound like I expect them to, even with the
two missing channels left. It's hard to find other examples of the
sweep function to test with.

I also fixed a mistake where I forgot to implement the length timer
right/at all. I'm not 100% sure it's correct now either, because the
specs are a little sparse on it, but I'm not entirely sure how it
could be wrong.

The out-of-sync problem is still there, and it's bad enough that I
think I need to fix that first. It may be that the sample rate is
lower than what sdl2 reports, or (less likely) that gaps in the sound
between frames are adding up since we're not calculating ahead, but
I'd think that would be audible.

* Fixed the sweep channel problem

The sweep channel didn't sound right; it turned out that it was
because I was triggering in setreg() before finishing the register
writes, so the upper 3 bits of the period were being discarded.

Still not sure what to do about synchronization. The emulator doesn't
run the 0x400000 clocks per second I expected, so that's a problem,
but it's not enough to explain the size of the desyncing anyway. I'll
probably just have to change how it's clocked entirely and accomodate
faster/slower emulation at the same time.

Co-authored-by: Mads Ynddal <mads-github@ynddal.dk>