Redo the memory window with Terminus bitmap font, removing TTF. Needs Cython.
This is a somewhat big commit. All refs to TTF were deleted, as well
as the various SDL2 pxd files that we're not using. I added a /font/
folder which contains the compressed font info, a short README, and
the font's license. Embedding/bundling the font is allowed under the
OFL, so we should be fine in that respect.

The portion of the font I included was just the 256 characters needed
to represent the IBM-437 codepage. It's an extension of ASCII that
includes a lot of useful lines for laying out windows and stuff, as
you'll see. Python works with this just fine, so hopefully that's not
an issue in the future. I have a script somewhere that I used to
generate the font string but I didn't think it's worth adding to the
repo. It wouldn't be too hard to replace if needed.

The memory window works well and looks good, but I don't have Cython
working yet with it. I used a lot of SDL calls that may require some
extra work. It works okay with PyPy but it is VERY SLOW. There may be
some optimizations needed to make this acceptable, or at least allow
disabling of the memory window.

Like the original memory window, scrolling is done with J/K, but now
faster scrolling uses Shift+J/K instead of the actual Shift keys. It
was surprisingly easy.

Hopefully precommit won't touch this commit because I think there are
config changes in master that this would conflict with on a lot of
files. I already had to undo that once.