Add function to read multibyte data
Add a function in the game wrapper to read multibyte data (in this case a pointer). this is almost certainly a feature that should either go in PyBoy proper or not exists at all (just manually reading consecutive bytes and adding them was the previous way), but it's being included here for now for ease of use and can be removed later if deemed necessary.