CHANGE: Added Set of Values
The purpose of these values is to allow the machine learning model to
gain information that it would not know of without it being complex.

Currently there is only access to the total number of items to provide
agents to some information in regards to their items.

The total number of pokemon in the party will allow the user to provide
rewards based on the number of pokemon the agent has gathered.

The time will provide the user with some insight as to how many in game
hours the model took to reach certain check points or flags found within
the game.

Added a battle state variable which will allow the user to either use
models that are specific to the battle scenarios within the game and the
game overworld.

Currently working on developing the flags for the badges. Have added it
on for now.