CHANGE: Reformat Time & Removed Code on Init
Removed the code that looked up the addresses when the game wrapper
class is initialized. This actually caused an error when using pyboy
through the command line.

After reading up and studying on bytes and how nintendo stores time, the
game_time variable was reformatted by using the bcd_to_dec and
_bcd_to_dec functions to extract the hours and minutes of in game time.

Made some changes in the comments to include more information. Excluded
the use of the badges until I become able to view the individual bits of
the byte which stores the badge information.